ENH: The filter outputs the weighted labels average
The filter is generic and outputs the weighted average for a label.
It checks for the number of labels present in the segmentations and
calculates the weighted average for each label present in the image.
If an output directory is specified, the outputs are written in the
directory using numbers starting at 1 (1.nrrd, 2.nrrd …).
If the output filenames are specified (NeoSeg pipeline), the output
images position from the filter will correspond to white = 1, gray = 2,
csf = 3.