ENH: Addition of FA to ABC pipeline
The WM extraction from FA is combined with the current weighted average
of the white matter.
In the case of ABC, the averages are written based in ascending order
of label value. For this purpose, the label value must be selected from
a combo box if the ABC pipeline is going to be used.
The combo box is named ‘comboBoxLabelValueFA’