ENH: Executing label reassign for a specific label.
The ‘reassign label’ can now be executed for any given label.
The algorithm computes the number of connected components per label. If
the number of voxels is below a threshold, then it checks for the
highest probability value. Two options are possible, check for the
highest probability voxel-by-voxel or compute the maximum probability
for the whole region.
The new label is assigned for the label with highest probability.